+++ src/utils/model.py
@@ -278,192 +278,3 @@ class Model():
                 y_pred = y_pred
             )
         )
-
-
-def finetune_model(
-        model_index: int,
-        X_train: np.array, 
-        y_train: np.array, 
-        X_val: np.array, 
-        y_val: np.array):
-    """
-    Iterate through all combinations of configs to fine-tune model
-    Args:
-        model_index (int): Index of the model
-        X_train (np.array): Array of BERT-tokenized post tokens from 
-            training set (X_train_tokenized)
-        y_train (np.array): Array of labels from training set (y_train)
-        X_val (np.array): Array of BERT-tokenized post tokens from 
-            training set (X_val_tokenized)
-        y_val (np.array): Array of labels from training set (y_val)
-    """
-
-    optimizer_names = [
-       "sgd", "rmsprop", "adadelta", "adagrad", 
-       "adam", "adamax", "adafactor", "nadam", "ftrl", "adamw"
-    ]
-    learning_rates = [1e-3, 1e-4, 1e-5, 1e-6]
-
-    for optimizer_name in optimizer_names:
-        for learning_rate in learning_rates:
-            # clear session before training a new model
-            tf.keras.backend.clear_session()
-
-            # build a model
-            model = Model(
-                model_index = model_index,
-                trainable = (model_index == 2),
-                optimizer_name = optimizer_name,
-                learning_rate = learning_rate,
-            )
-                
-            # finetune model
-            model_hist = model.finetune(
-                X_train = X_train,
-                y_train = y_train,
-                X_val = X_val,
-                y_val = y_val,
-            )
-
-
-def evaluate_tf_model(
-      filename: str,
-      X_test: np.array,
-      y_test: np.array):
-    """
-    After running all experiments, pick the best iterations for models 1 and 2 
-    based on validation scores and then run evaluate using test set
-
-    Args:
-        filename (str): Filename to the saved model weights
-        X_test (np.array): Features for test set
-        y_test (np.array): Labels for test set
-    """
-    tf.keras.backend.clear_session()
-
-    model = Model(weights_filename = filename)
-    model.evaluate(X_test = X_test, y_test = y_test)
-
-
-def evaluate_pt_model(
-        X_test: np.array,
-        y_test: np.array,
-        pretrained_model_name: str = "Hate-speech-CNERG/bert-base-uncased-hatexplain"):
-    """
-    After running all experiments, pick the best iterations for model 3 
-    based on validation scores and then run evaluate using test set
-
-    Args:
-        X_test (np.array): Features for test set
-        y_test (np.array): Labels for test set
-        pretrained_model_name (str): Name of pretrained model
-    """
-
-    pipe = transformers.pipeline(
-        "text-classification",
-        model = pretrained_model_name)
-    
-    y_pred = np.array([
-        int(prediction["label"] != "normal") for prediction in pipe(X_test)
-    ])
-
-    if isinstance(y_test, list):
-        y_test = np.array(y_test)
-
-    print(
-        classification_report(
-            y_true = y_test,
-            y_pred = y_pred
-        )
-    )
-
-
-def save_tf_hidden_states(
-        model_index: int,
-        weights_filename: str, 
-        X_test_tokenized: list[str],
-        folderpath: str = "../data/hidden_states/"):
-    """
-    Save hidden states as numpy arrays for models built using tensorflow under
-    the folderpath directory.
-
-    Args:
-        model_index (int): Index of the model
-        weights_filename (str): Filename of the saved model weights
-        X_test_tokenized (list[str]): List of tokenized subword tokens to use
-            in extracting the hidden states
-    """
-    # load models
-    model = Model(
-        weights_filename = weights_filename,
-        output_hidden_states = True,
-    )
-
-    model_index = re.findall(R"(model\_\d+)_.*", weights_filename)[0]
-    filename_prefix = F"{model_index}_hidden_states"
-
-    for test_set_idx, text in enumerate(X_test_tokenized):
-
-        logging.debug(F"Processing test_set_idx={test_set_idx}")
-
-        # generate prediction
-        result = model.model_object.predict(text)
-
-        # save hidden states
-        filepath = os.path.join(
-            folderpath, 
-            F"{filename_prefix}_test_set_{test_set_idx}_hidden_states.npy"
-        )
-        with open(filepath, "wb") as fp:
-            np.save(fp, np.array(result.hidden_states))
-
-
-def save_pt_hidden_states(
-        model_index: int, 
-        pretrained_model_name: str, 
-        X_test: list[str],
-        folderpath: str = "../data/hidden_states/"):
-    """
-    Save hidden states as numpy arrays for models built using pytorch under
-    the directory ../data/hidden_states/.
-
-    Args:
-        model_index (int): Index of the model
-        pretrained_model_name (str): Name of pretrained tokenizer
-        X_test (list[str]): List of post contents to use in extracting the 
-            hidden states
-    """
-    # set up pretrained tokenizer
-    tokenizer = transformers.AutoTokenizer.from_pretrained(
-        pretrained_model_name,
-        max_length = 200,
-        truncation = True,
-        padding = 'max_length'
-    )
-
-    # set up pretrained model
-    model = transformers.AutoModelForSequenceClassification.from_pretrained(
-        pretrained_model_name,
-        output_hidden_states = True
-    )
-
-    filename_prefix = F"model_{model_index}_hidden_states"
-
-    for test_set_idx, text in enumerate(X_test):
-
-        logging.debug(F"Processing test_set_idx={test_set_idx}")
-
-        # tokenizer text
-        inputs = tokenizer(text, return_tensors = "pt")
-
-        # generate prediction
-        with torch.no_grad():
-            result = model(**inputs)
-
-        # save hidden states
-        filepath = os.path.join(
-            folderpath,
-            F"{filename_prefix}_test_set_{test_set_idx}_hidden_states.npy"
-        )
-        with open(filepath, "wb") as fp:
-            np.save(fp, np.array([h.numpy() for h in result.hidden_states]))
